requirements.txt
flask==3.0.0
python-docx==1.1.0
ollama==0.3.3
werkzeug==3.0.1
